backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.115.12",
    "google-adk>=1.2.1",
    "google-genai>=1.19.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "mem0ai>=0.1.104",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
